Moved LoraHook stuff into conditioning.py from utilts_motion.py
