+++ animatediff/sample_settings.py
@@ -9,9 +9,10 @@ from comfy.model_patcher import ModelPatcher
 from comfy.model_base import BaseModel
 
 from . import freeinit
+from .conditioning import LoraHookMode
 from .context import ContextOptions, ContextOptionsGroup
 from .utils_model import SigmaSchedule
-from .utils_motion import extend_to_batch_size, get_sorted_list_via_attr, prepare_mask_batch, LoraHookMode
+from .utils_motion import extend_to_batch_size, get_sorted_list_via_attr, prepare_mask_batch
 from .logger import logger
